Migrate `tests/unit/modules/test_cmdmod.py` to pytest
